# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tab {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  workspaceId: number | null;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  groupName?: string | null;
  /** @nullable */
  groupColor?: string | null;
  isActive: boolean;
  isSleeping: boolean;
  isPinned: boolean;
  isIncognito: boolean;
  /** @nullable */
  memoryMb?: number | null;
  createdAt: string;
}

export interface TabInput {
  url: string;
  title?: string;
  /** @nullable */
  workspaceId?: number | null;
  isIncognito?: boolean;
  isPinned?: boolean;
}

export interface TabUpdate {
  title?: string;
  url?: string;
  isActive?: boolean;
  isSleeping?: boolean;
  isPinned?: boolean;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  workspaceId?: number | null;
}

export interface TabStats {
  total: number;
  sleeping: number;
  pinned: number;
  incognito: number;
  grouped: number;
  totalMemoryMb: number;
}

export interface Workspace {
  id: number;
  name: string;
  icon: string;
  color: string;
  isActive: boolean;
  tabCount: number;
  createdAt: string;
}

export interface WorkspaceInput {
  name: string;
  icon: string;
  color: string;
}

export interface WorkspaceUpdate {
  name?: string;
  icon?: string;
  color?: string;
  isActive?: boolean;
}

export interface Bookmark {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  folder?: string | null;
  /** @nullable */
  workspaceId?: number | null;
  createdAt: string;
}

export interface BookmarkInput {
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  folder?: string | null;
  /** @nullable */
  workspaceId?: number | null;
}

export interface HistoryEntry {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
  visitCount?: number;
  visitedAt: string;
}

export interface HistoryInput {
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
}

export interface TopSite {
  url: string;
  title: string;
  visitCount: number;
  /** @nullable */
  favicon: string | null;
}

export interface AiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  model?: string | null;
  createdAt: string;
}

export interface AiMessageInput {
  content: string;
  /** @nullable */
  conversationId?: number | null;
  /** @nullable */
  context?: string | null;
}

export interface Conversation {
  id: number;
  title: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SmartSuggestion {
  id: number;
  type: string;
  title: string;
  url: string;
  reason: string;
  /** @nullable */
  favicon?: string | null;
}

export interface Download {
  id: number;
  filename: string;
  url: string;
  status: string;
  /** @nullable */
  sizeBytes: number | null;
  downloadedBytes: number;
  createdAt: string;
}

export interface DownloadInput {
  url: string;
  filename: string;
}

export interface DashboardOverview {
  totalTabs: number;
  sleepingTabs: number;
  trackersBlocked: number;
  memoryUsedMb: number;
  dataSavedMb: number;
  aiUsesToday: number;
  workspaceCount: number;
  bookmarkCount: number;
  historyCount: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  /** @nullable */
  url?: string | null;
  timestamp: string;
}

export interface SyncStatus {
  /** @nullable */
  lastSyncAt: string | null;
  deviceCount: number;
  tabsSynced: boolean;
  bookmarksSynced: boolean;
  historySynced: boolean;
  settingsSynced: boolean;
}

export interface SyncPushInput {
  includeTabs?: boolean;
  includeBookmarks?: boolean;
  includeHistory?: boolean;
  includeSettings?: boolean;
}