# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
servers:
  - url: /api
    description: EoN Browser API Gateway
...
paths:
...
  /tabs:
    get:
      operationId: listTabs
      tags: [tabs]
      summary: List all open tabs
      responses:
        "200":
          description: List of tabs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tab"
    post:
      operationId: createTab
      tags: [tabs]
      summary: Open a new tab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TabInput"
      responses:
        "201":
          description: Tab created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tab"
  /tabs/{id}:
    get:
      operationId: getTab
      tags: [tabs]
      summary: Get a tab by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tab found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tab"
    patch:
      operationId: updateTab
      tags: [tabs]
      summary: Update tab properties
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TabUpdate"
      responses:
        "200":
          description: Tab updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tab"
    delete:
      operationId: closeTab
      tags: [tabs]
      summary: Close a tab
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tab closed
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── BOOKMARKS ─────────────────────────────────────────────────────────────
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List all bookmarks
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Add a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Tab:
      type: object
      required: [id, title, url, workspaceId, isActive, isSleeping, isPinned, isIncognito, createdAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        url:         { type: string }
        favicon:     { type: ["string", "null"] }
        workspaceId: { type: ["integer", "null"] }
        groupId:     { type: ["integer", "null"] }
        groupName:   { type: ["string", "null"] }
        groupColor:  { type: ["string", "null"] }
        isActive:    { type: boolean }
        isSleeping:  { type: boolean }
        isPinned:    { type: boolean }
        isIncognito: { type: boolean }
        memoryMb:    { type: ["number", "null"] }
        createdAt:   { type: string }
    TabInput:
      type: object
      required: [url]
      properties:
        url:         { type: string }
        title:       { type: string }
        workspaceId: { type: ["integer", "null"] }
        isIncognito: { type: boolean }
        isPinned:    { type: boolean }
    TabUpdate:
      type: object
      properties:
        title:       { type: string }
        url:         { type: string }
        isActive:    { type: boolean }
        isSleeping:  { type: boolean }
        isPinned:    { type: boolean }
        groupId:     { type: ["integer", "null"] }
        workspaceId: { type: ["integer", "null"] }
...
    Workspace:
      type: object
      required: [id, name, icon, color, isActive, tabCount, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        icon:      { type: string }
        color:     { type: string }
        isActive:  { type: boolean }
        tabCount:  { type: integer }
        createdAt: { type: string }
    WorkspaceInput:
      type: object
      required: [name, icon, color]
      properties:
        name:  { type: string }
        icon:  { type: string }
        color: { type: string }
    WorkspaceUpdate:
      type: object
      properties:
        name:     { type: string }
        icon:     { type: string }
        color:    { type: string }
        isActive: { type: boolean }

    # ── BOOKMARKS ────────────────────────────────────────────────────────────
    Bookmark:
      type: object
      required: [id, title, url, createdAt]
      properties:
        id:          { type: integer }
        title:       { type: string }
        url:         { type: string }
        favicon:     { type: ["string", "null"] }
        folder:      { type: ["string", "null"] }
        workspaceId: { type: ["integer", "null"] }
        createdAt:   { type: string }
    BookmarkInput:
      type: object
      required: [title, url]
      properties:
        title:       { type: string }
        url:         { type: string }
        favicon:     { type: ["string", "null"] }
        folder:      { type: ["string", "null"] }
        workspaceId: { type: ["integer", "null"] }

    # ── HISTORY ──────────────────────────────────────────────────────────────
...