# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
 * OpenAPI spec version: 1.0.0
 */

export interface Bookmark {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  folder?: string | null;
  /** @nullable */
  workspaceId?: number | null;
  createdAt: string;
}