# lib/api-zod/src/generated/types/tab.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
 * OpenAPI spec version: 1.0.0
 */

export interface Tab {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  favicon?: string | null;
  /** @nullable */
  workspaceId: number | null;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  groupName?: string | null;
  /** @nullable */
  groupColor?: string | null;
  isActive: boolean;
  isSleeping: boolean;
  isPinned: boolean;
  isIncognito: boolean;
  /** @nullable */
  memoryMb?: number | null;
  createdAt: string;
}