# lib/api-zod/src/generated/types/tabInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
 * OpenAPI spec version: 1.0.0
 */

export interface TabInput {
  url: string;
  title?: string;
  /** @nullable */
  workspaceId?: number | null;
  isIncognito?: boolean;
  isPinned?: boolean;
}