# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
 * OpenAPI spec version: 1.0.0
 */

export interface Workspace {
  id: number;
  name: string;
  icon: string;
  color: string;
  isActive: boolean;
  tabCount: number;
  createdAt: string;
}