# lib/api-zod/src/generated/types/workspaceUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EoN Browser backend API — tabs, workspaces, bookmarks, history, AI intelligence, sync, and security systems
 * OpenAPI spec version: 1.0.0
 */

export interface WorkspaceUpdate {
  name?: string;
  icon?: string;
  color?: string;
  isActive?: boolean;
}